+++ dbt/include/sqlserver/macros/materializations/models/incremental/incremental.sql
@@ -130,6 +130,10 @@
 
   {{ run_hooks(post_hooks, inside_transaction=False) }}
 
+  {# Build any ONLINE/RESUMABLE indexes now the model transaction has committed;
+     they cannot run inside a transaction. No-op when none are configured. #}
+  {% do sqlserver__create_indexes_post_commit(target_relation) %}
+
   {{ return({'relations': [target_relation]}) }}
 
 {%- endmaterialization %}